finished fourth and fifth challenges
