+++ package.json
@@ -19,6 +19,7 @@
     "babel-plugin-styled-components": "^2.1.4",
     "dompurify": "^3.0.5",
     "highlight.js": "^11.8.0",
+    "localforage": "^1.10.0",
     "marked": "^5.1.1",
     "normalize.css": "^8.0.1",
     "react": "^18.2.0",
